reorganize ResultSet
