+++ src/test/java/net/snowflake/client/jdbc/ResultSetArrowLatestIT.java
@@ -4,8 +4,8 @@ import net.snowflake.client.category.TestCategoryArrow;
 import org.junit.experimental.categories.Category;
 
 @Category(TestCategoryArrow.class)
-public class ResultSetArrowIT extends ResultSetIT {
-  public ResultSetArrowIT() {
+public class ResultSetArrowLatestIT extends ResultSetLatestIT {
+  public ResultSetArrowLatestIT() {
     super("arrow");
   }
 }
